Moving Game and Lobby activities and fragments to their own package.
