+++ mobile/Cards/app/src/main/res/layout/activity_lobby.xml
@@ -1,7 +1,7 @@
 <android.support.v4.widget.DrawerLayout
     xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="cards.seniordesign.com.cards.Lobby"
+    tools:context="cards.seniordesign.com.cards.lobby.Lobby"
     style="@style/ActivityBase"
     android:id="@+id/lobby_drawer">
