+++ mobile/Cards/app/src/main/res/layout/fragment_list_rooms.xml
@@ -1,7 +1,7 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools" android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context="cards.seniordesign.com.cards.ListRoomsFragment">
+    tools:context="cards.seniordesign.com.cards.lobby.ListRoomsFragment">
 
     <ScrollView
         android:layout_width="match_parent"
