+++ mobile/Cards/app/src/main/res/menu/menu_game.xml
@@ -1,6 +1,6 @@
 <menu xmlns:android="http://schemas.android.com/apk/res/android"
     xmlns:tools="http://schemas.android.com/tools"
-    tools:context="cards.seniordesign.com.cards.Game">
+    tools:context="cards.seniordesign.com.cards.game.Game">
     <item android:id="@+id/action_settings" android:title="@string/action_settings"
         android:orderInCategory="100" android:showAsAction="never" />
 </menu>
